api-gateway/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0